+++ app/Http/Controllers/Admin/AccommodationController.php
@@ -6,17 +6,15 @@ use App\Accommodation;
 use App\AccommodationPhoto;
 use App\AccommodationType;
 use App\AccommodationsAvailabilityIntervals;
-use App\Country;
 use App\FacilityType;
-use App\HelpRequestAccommodationDetail;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AccommodationRequest;
+use App\Services\AccommodationService;
 use App\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
@@ -117,17 +115,10 @@ class AccommodationController extends Controller
         /** @var User $user */
         $user = Auth::user();
 
-        return view('admin.accommodation-edit')
-            ->with('user', $user)
-            ->with('accommodation', $accommodation)
-            ->with('types', AccommodationType::all())
-            ->with('ownershipTypes', Accommodation::getOwnershipTypes())
-            ->with('generalFacilities', FacilityType::where('type', '=', FacilityType::TYPE_GENERAL)->get())
-            ->with('specialFacilities', FacilityType::where('type', '=', FacilityType::TYPE_SPECIAL)->get())
-            ->with('otherFacilities', FacilityType::where('type', '=', FacilityType::TYPE_OTHER)->first())
-            ->with('availabilityIntervals', $accommodation->availabilityIntervals()->get())
-            ->with('countries', Country::all())
-            ->with('photoData', $this->getPhotoData($accommodation));
+        $accService = new AccommodationService();
+
+        return $accService->viewEditAccommodation($user, $accommodation, 'admin.accommodation-edit');
+
     }
 
     /**
@@ -268,15 +259,9 @@ class AccommodationController extends Controller
             abort(404);
         }
 
-        return view('admin.accommodation-add')
-            ->with('user', $user)
-            ->with('types', AccommodationType::all())
-            ->with('ownershipTypes', Accommodation::getOwnershipTypes())
-            ->with('generalFacilities', FacilityType::where('type', '=', FacilityType::TYPE_GENERAL)->get())
-            ->with('specialFacilities', FacilityType::where('type', '=', FacilityType::TYPE_SPECIAL)->get())
-            ->with('otherFacilities', FacilityType::where('type', '=', FacilityType::TYPE_OTHER)->first())
-            ->with('countries', Country::all())
-            ->with('counties', County::all());
+        $accService = new AccommodationService();
+
+        return $accService->viewAddAccommodation($user, 'admin.accommodation-add');
     }
 
     /**
@@ -293,96 +278,15 @@ class AccommodationController extends Controller
             abort(400);
         }
 
-        DB::beginTransaction();
-
-        $accommodation = new Accommodation();
-        $accommodation->user_id = $userId;
-        $accommodation->accommodation_type_id = $request->get('type');
-        $accommodation->ownership_type = $request->get('ownership');
-        $accommodation->is_fully_available = ('fully' == $request->get('property_availability'));
-        $accommodation->max_guests = $request->get('max_guests');
-        $accommodation->available_rooms = $request->get('available_rooms');
-        $accommodation->available_beds = $request->get('available_beds');
-        $accommodation->available_bathrooms = $request->get('available_bathrooms');
-        $accommodation->is_kitchen_available = ('yes' == $request->get('allow_kitchen'));
-        $accommodation->is_parking_available = ('yes' == $request->get('allow_parking'));
-        $accommodation->is_smoking_allowed = ('yes' == $request->get('allow_smoking'));
-        $accommodation->is_pet_allowed = ('yes' == $request->get('allow_pets'));
-        $accommodation->description = $request->get('description');
-        $accommodation->address_country_id = (int)$request->get('country');
-        $accommodation->address_county_id = (int)$request->get('county_id');
-        $accommodation->address_city = $request->get('city');
-        $accommodation->address_street = $request->get('street');
-        $accommodation->address_building = $request->get('building');
-        $accommodation->address_entry = $request->get('entrance');
-        $accommodation->address_apartment = $request->get('apartment');
-        $accommodation->address_floor = $request->get('floor');
-        $accommodation->address_postal_code = $request->get('postal_code');
-        $accommodation->other_rules = $request->get('other_rules');
-        $accommodation->transport_subway_distance = $request->get('transport_subway_distance');
-        $accommodation->transport_bus_distance = $request->get('transport_bus_distance');
-        $accommodation->transport_railway_distance = $request->get('transport_railway_distance');
-        $accommodation->transport_other_details = $request->get('transport_other_details');
-        $accommodation->save();
-
-        if ($request->has('general_facility')) {
-            foreach ($request->get('general_facility') as $key => $value) {
-                $accommodation->accommodationfacilitytypes()->attach($value);
-            }
+        try
+        {
+            $accService = new AccommodationService();
+            $accommodation = $accService->createAccommodation($request, $user);
         }
-
-        if ($request->has('special_facility')) {
-            foreach ($request->get('special_facility') as $key => $value) {
-                $accommodation->accommodationfacilitytypes()->attach($value);
-            }
-        }
-
-        if ($request->has('other_facilities') && !is_null($request->get('other_facilities'))) {
-            /** @var FacilityType|null $otherFacilityType */
-            $otherFacilityType = FacilityType::where('type', '=', FacilityType::TYPE_OTHER)->first();
-
-            if (!empty($otherFacilityType)) {
-                $accommodation->accommodationfacilitytypes()->attach($otherFacilityType->id, ['message' => $request->get('other_facilities')]);
-            }
-        }
-
-        if ($request->has("available") && is_array($request->get("available"))) {
-            foreach ($request->get("available") as $key => $value) {
-                $accommodationsAvailabilityInterval = new AccommodationsAvailabilityIntervals();
-                $accommodationsAvailabilityInterval->accommodation_id = $accommodation->id;
-                $accommodationsAvailabilityInterval->from_date = $request->get("available")[$key]['from'];
-                $accommodationsAvailabilityInterval->to_date = $request->get("available")[$key]['to'];
-                $accommodationsAvailabilityInterval->save();
-            }
-        }
-
-        try {
-            if (!empty($request->file('photos'))) {
-                /** @var UploadedFile $file */
-                foreach ($request->file('photos') as $file) {
-                    $fileName = sha1((string)microtime() . $file->getClientOriginalName()) . $file->getClientOriginalExtension();
-
-                    /** @var string $path */
-                    $path = Storage::disk('private')->putFile('accommodation/' . $accommodation->id . '/' . $fileName, $file);
-
-                    $accommodationPhoto = new AccommodationPhoto();
-                    $accommodationPhoto->accommodation_id = $accommodation->id;
-                    $accommodationPhoto->name = $file->getClientOriginalName();
-                    $accommodationPhoto->identifier = sha1($path);
-                    $accommodationPhoto->path = $path;
-                    $accommodationPhoto->size = $file->getSize();
-                    $accommodationPhoto->extension = '.' . $file->getClientOriginalExtension();
-                    $accommodationPhoto->type = $file->getClientMimeType();
-                    $accommodationPhoto->save();
-                }
-            }
-        } catch (\Throwable $throwable) {
-            DB::rollBack();
-
+        catch (\Throwable $throwable)
+        {
             return Redirect::back()->withInput()->withErrors(['photos' => $throwable->getMessage()]);
         }
-
-        DB::commit();
 
         return redirect()
             ->route('admin.host-detail', $accommodation->user_id)
