+++ resources/views/frontend/host/add-accommodation.blade.php
@@ -1,3 +1,3 @@
 @extends('layouts.app')
 
-@include('host.add-accommodation-base')
+@include('host.add-accommodation-base', ['formRoute' => route('get-involved-save-accommodation')])
