+++ resources/views/host/add-accommodation.blade.php
@@ -1,3 +1,3 @@
 @extends('layouts.admin')
 
-@include('host.add-accommodation-base')
+@include('host.add-accommodation-base', ['formRoute' => route('host.create-accommodation')])
